# Use character array for utfstring