--- a/src/UtfString.php
+++ b/src/UtfString.php
@@ -8,12 +8,10 @@ use ArrayAccess;
 use Exception;
 use Stringable;
 
+use function count;
+use function implode;
 use function mb_check_encoding;
-use function mb_strlen;
-use function mb_substr;
-use function ord;
-use function strlen;
-use function substr;
+use function mb_str_split;
 
 /**
  * Implementation for UTF-8 strings.
@@ -32,44 +30,19 @@ use function substr;
 class UtfString implements ArrayAccess, Stringable
 {
     /**
-     * The raw, multi-byte string.
-     */
-    public string $str = '';
-
-    /**
-     * The index of current byte.
+     * The multi-byte characters.
      *
-     * For ASCII strings, the byte index is equal to the character index.
+     * @var list<string>
      */
-    public int $byteIdx = 0;
-
-    /**
-     * The index of current character.
-     *
-     * For non-ASCII strings, some characters occupy more than one byte and
-     * the character index will have a lower value than the byte index.
-     */
-    public int $charIdx = 0;
-
-    /**
-     * The length of the string (in bytes).
-     */
-    public int $byteLen = 0;
-
-    /**
-     * The length of the string (in characters).
-     */
-    public int $charLen = 0;
+    public array $characters;
 
     /** @param string $str the string */
     public function __construct(string $str)
     {
-        $this->str = $str;
-        $this->byteLen = mb_strlen($str, '8bit');
-        if (! mb_check_encoding($str, 'UTF-8')) {
-            $this->charLen = 0;
+        if (mb_check_encoding($str, 'UTF-8')) {
+            $this->characters = mb_str_split($str, 1, 'UTF-8');
         } else {
-            $this->charLen = mb_strlen($str, 'UTF-8');
+            $this->characters = [];
         }
     }
 
@@ -80,7 +53,7 @@ class UtfString implements ArrayAccess, Stringable
      */
     public function offsetExists(mixed $offset): bool
     {
-        return ($offset >= 0) && ($offset < $this->charLen);
+        return $offset >= 0 && $offset < count($this->characters);
     }
 
     /**
@@ -90,37 +63,7 @@ class UtfString implements ArrayAccess, Stringable
      */
     public function offsetGet(mixed $offset): string|null
     {
-        // This function moves the internal byte and character pointer to the requested offset.
-        // This function is part of hot code so the aim is to do the following
-        // operations as efficiently as possible.
-        // UTF-8 character encoding is a variable length encoding that encodes Unicode
-        // characters in 1-4 bytes. Thus we fetch 4 bytes from the current offset and then use mb_substr
-        // to get the first UTF-8 character in it. We then use strlen to get the character's size in bytes.
-        if (($offset < 0) || ($offset >= $this->charLen)) {
-            return null;
-        }
-
-        $delta = $offset - $this->charIdx;
-
-        if ($delta > 0) {
-            // Fast forwarding.
-            $this->byteIdx += strlen(mb_substr(substr($this->str, $this->byteIdx, 4 * $delta), 0, $delta));
-            $this->charIdx += $delta;
-        } elseif ($delta < 0) {
-            // Rewinding.
-            while ($delta++ < 0) {
-                // We rewind byte by byte and only count characters that are not continuation bytes,
-                // i.e. ASCII characters and first octets of multibyte characters
-                do {
-                    $byte = ord($this->str[--$this->byteIdx]);
-                } while (($byte >= 128) && ($byte < 192));
-
-                --$this->charIdx;
-            }
-        }
-
-        // Fetch the first Unicode character within the next 4 bytes in the string.
-        return mb_substr(substr($this->str, $this->byteIdx, 4), 0, 1);
+        return $this->characters[$offset] ?? null;
     }
 
     /**
@@ -153,7 +96,7 @@ class UtfString implements ArrayAccess, Stringable
      */
     public function length(): int
     {
-        return $this->charLen;
+        return count($this->characters);
     }
 
     /**
@@ -161,6 +104,6 @@ class UtfString implements ArrayAccess, Stringable
      */
     public function __toString(): string
     {
-        return $this->str;
+        return implode('', $this->characters);
     }
 }