--- a/tests/data/lexer/lexUtf8.out
+++ b/tests/data/lexer/lexUtf8.out
@@ -6,11 +6,7 @@
         "errors": [],
         "str": {
             "@type": "PhpMyAdmin\\SqlParser\\UtfString",
-            "str": "select * from école",
-            "byteIdx": 19,
-            "charIdx": 18,
-            "byteLen": 20,
-            "charLen": 19
+            "str": "select * from école"
         },
         "len": 19,
         "last": 19,