Benchmarks indicate a nice performance gain, too.

Old:
| benchmark      | subject                               | set | revs | its | mem_peak | mode     | rstdev |
|----------------|---------------------------------------|-----|------|-----|----------|----------|--------|
| UtfStringBench | benchBuildUtfString                   |     | 4    | 20  | 1.910mb  | 34.245ms | ±1.33% |
| UtfStringBench | benchUtfStringRandomAccessWithUnicode |     | 4    | 20  | 1.910mb  | 78.063μs | ±9.44% |

New:
| benchmark      | subject                               | set | revs | its | mem_peak | mode     | rstdev |
|----------------|---------------------------------------|-----|------|-----|----------|----------|--------|
| UtfStringBench | benchBuildUtfString                   |     | 4    | 20  | 2.513mb  | 14.636ms | ±2.37% |
| UtfStringBench | benchUtfStringRandomAccessWithUnicode |     | 4    | 20  | 1.910mb  | 22.511μs | ±3.69% |