@williamdes 2024-08-26
Maybe we should target 5.10.x ?
Or keep 6.0 ?